partialize sync execution component
